ifc6410: dragonboard410c: add mesa-driver-msm is added in wayland images
So far, mesa-driver-msm is included only if/when X11 is used since we included
it through XSERVER variable. As such images with Wayland/Weston do not get it,
and will fail to use the GPU.

This commit adds mesa-driver-msm as a machine RRECOMMENDS, if opengl is enabled
in DISTRO_FEATURES. This is still not quite ideal, since we don't need it in all
images, e.g. console only image should not have it... So we still need to fix it
properly, but in the mean time this is now 'less worse'..

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>