linux: allow the kernel recipe to use defconfig from git tree
Instead of requiring the defconfig file to be in SRC_URI, we leave the choice to
the kernel recipe whether it has a local defconfif file, or if it uses a config
file from the git tree directly, which can be specific with an absolute path in
KERNEL_DEFCONFIG variable.

If both the variable is set and the local file defconfig exists in SRC_URI, then
the local file is used.

This commit only adds a new feature, but none of the kernel recipe uses it for now.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>